add m-to-m relation between scientist and positive traits tables, fix types in tables

+++ src/main/resources/db/migration/create_tables/V012__create_m_to_m_table_scientists_achievements.sql
@@ -1,6 +1,6 @@
 CREATE TABLE scientists_achievements
 (
-    scientist_id   INT REFERENCES scientists (scientist_id),
-    achievement_id INT REFERENCES achievements (achievement_id),
+    scientist_id   BIGSERIAL REFERENCES scientists (scientist_id),
+    achievement_id SERIAL REFERENCES achievements (achievement_id),
     PRIMARY KEY (scientist_id, achievement_id)
 );
